Overly complex day 4 part 1 solution
Either the difficulty went way up or I solved this in a silly way. Might
be one of those examples where the "naive" approach isn't actually that
naive. Oh well.

One direction I almost foolishly took was to try and reverse each of the
strings inside the rows/columns/diagonals to look for "XMAS". It's much
simpler to just check for both "XMAS" and "SAMX" at the same time vs.
reversing each of the strings and iterating through 'em twice.