Day 5 Part 2: Naive solution
Uses the same `isValidUpdate` to check if the update is valid, then just
does a straightforward hack to turn the rules into a comparison function
for ordering.

I started trying to do something a bit more clever than just treating
the rules as an ordering; I was going to try and return all the out-of-
order items in a given update. But then I couldn't figure out a good
strategy to get them *back* in order; with the non-linear "isValidUpate"
I couldn't figure out a better way than just "swap, check if valid,
identify out-of order, swap, check if valid, swap out of orders...",
which didn't seem like a very satisfying approach.