Day 6 Part 1
Think this is pretty close to space optimal; was originally going to
generate an entire "next map" for each step but realized using the map
storage as a tracker was actually pretty straightforward and
space-efficient.

Think this is close to time-optimized as well for a naive solution; of
course there may be more clever ways to track the direction and speed
up the search for the next direction change.