Day 8 Part 1
Pretty straightforward brute force for Part 1. N is really small for
this problem and this is O(len(antennas)^2). I don't think it could be
much smaller than this; maybe you could reduce constants by calculating
antinodes while finding antennas, but this is already as fast in clock
time as just processing the input so I don't think there's a lot of fat.