Day 2 Part 2 - super copy-paste
Gonna clean this up, but this is the basic solution. Only snag I hit was
checking the set length as my exit condition (as a proxy for whether we
added to the set), but forgetting that even if the set didn't grow, we
still might not be off the map yet.