Day 9 Part 1 - Gahhhhh
Spent way too long putting `generated` in a string and forgetting that
IDs could be 2 digits...cheated by looking at the below 😔
https://www.reddit.com/r/adventofcode/comments/1haeecm/its_been_years_of_debugging/

Oh well. There's definitely a more space-efficient way to do it by
storing the compactified representation of the generated string, but I'm
too lazy to figure it out.