Day 10 Part 2: Super straightforward
I actually already implemented this logic when initially attempting to
Part 1 due to some confusion. I was able to recover the initial attempt
with Cmd+z, and then it was straightforward to convert it to the shorter
approach of Part 1. There's probably a way to dedupe some of this code
a tad more.