Bam, figured out the Part 1 refactor. That was satisfying
Use the same "do I need a fence?" properties as before to count up the
perimeter (off the map or next option doesn't match), but counting it up
while finding the region makes the code a heck of a lot shorter.

Probably part of the reason I couldn't unlock this earlier is because I
wasn't representing `region.rows` in as easily-indexable way, so I
wasn't able to make the important nuance between `seen` (global-level
memo to avoid processing the same regions 2x) and `region.rows` (local
memo to avoid processing the same plot 2x when finding a region).