Day 12 Part 2: Uuuuugly code with wasted space and copy-pasta
But it works in 50ms, so 🤷‍♂️. This one was tricky to figure out, at first
I was even having trouble figuring out an algorithm to count the sides
manually. I eventually realized I was doing it by first scanning top to
bottom, looking for "breaks" in the top or bottom fence of a row to add
another side, then going left to right and looking for a "break" in the
right or left fence of a column.

I already had the rows of the Region easily accessible, so it was quick
to test out the approach top to bottom and see if it'd work. I stored
the sides of each square that had a fence, figuring it was only a 4x
space increase on a pretty small problem in memory, and then just wrote
the cases for what a "break" was. Once I saw that worked, I decided to
store the columns too and copy-pasta the code.

I'm working with ChatGPT to try and figure out a way to optimize, but
for now I'm OK with this.