A ChatGPT-aided refactor
I asked ChatGPT for some help with both reducing the duplicated storage
in the DS, as well as fixing up cleanSides. It did a great job with the
DS:
> cells: A dictionary where the key is a tuple (x, y) (the coordinates
> of a square), and the value is a set containing the directions of the
> fences (UP, DOWN, LEFT, RIGHT) for that cell.

I thought using a set as the value of the dict was overkill since the
max length is 4, but otherwise it was helpful.

The main insight for cleaning up countSides was that while for human
me to actually count the sides I have to parse through top-to-bottom &
left-to-right, in reality the `if` statements being checked only
depended on the current `x, y` and needed to be run at least once for
every plot in the region, so just iterating through all the plots in
the region and smushing the conditions together has the same result.