Day 13 Part 1: Naive simple approach
The problem wording gave a pretty good hint, but this is solvable in
O(N) time because the limit on button presses is 100. So you just
calculate (end - cheap button presses, cost) and (start + pricey button
presses, cost) for 100 button presses. If (end - cheap) == (0,0) at any
point, you count that as a potential cost, and if (start+pricey) = prize
you count that as a potential cost.

The "tricky" bit is the intuition that you don't actually need to
consider every possible *path*, you just need to consider the counts of
button presses, but this isn't that tricky since the problem statement
strongly hints at it.

🍏 time python3 day13.py pt1 input.txt
['day13.py', 'pt1', 'input.txt']
totalCost: 29522

________________________________________________________
Executed in  253.10 millis    fish           external
   usr time  237.95 millis   78.00 micros  237.87 millis
   sys time   13.80 millis  894.00 micros   12.91 millis