Day 13 Part 2: Finally (I hate floating point arithmetic)
Today was a real doozy. I got hung up on the idea that this was a single
equation (deltaA*Apresses + deltaB*Bpresses = prize) and spent too long
on a Diophantine equation rabbit hole before realizing it's actually
just a system of 2 linear equations w/ 2 unknowns, which is guaranteed
to have exactly 1 solution.

However, even after that eureka moment and solving the system of eqtns,
I spent an embarrassing amount of time on some edge cases
1. The solution only "counts" if x,y are +ive integers. I spent way too
   long trying different mechanisms to coax integer arithmetic, and then
   forgot that `int()` truncates, before settling on the simplest
   mechanism of double-checking the solution after `round()`-ing
2. If the 2 buttons have the same slope line on their path, you end up
   with effectively 1 equation with 2 unknowns that define a line of
   infinite solutions. But in that edge case, you simply take the button
   that's cheaper token cost/distance all the way to the end.

Worth the struggle though, it was so satisfying in the end:
🍏 time python3 day13.py pt2 input.txt
['day13.py', 'pt2', 'input.txt']
totalCost: <correct answer>

________________________________________________________
Executed in   44.97 millis    fish           external
   usr time   29.92 millis    0.09 millis   29.82 millis
   sys time   13.61 millis    1.09 millis   12.52 millis

🍏 time python3 day13.py pt1 input.txt
['day13.py', 'pt1', 'input.txt']
totalCost: <correct answer>

________________________________________________________
Executed in   36.18 millis    fish           external
   usr time   25.43 millis   37.00 micros   25.39 millis
   sys time    9.65 millis  529.00 micros    9.12 millis