Day 14 Part 2
Pretty straightforward. Since they didn't even give any hints as to the format of the tree, I figured it was better to just print out all the candidate trees ("candidate" defined as any map that has a run of 10+ straight  bots, since I figured the tree would consist of a line) and indices and find the tree by visual inspection.