Day 15 Part 2: Finally
I really struggled with figuring out how to move the boxes vertically.
Processing the input and moving the boxes horizontally were both
straightforward, just required some minimal thinking.

But moving the boxes vertically was a struggle. I thought I had it after
testing on the smallest input/less small input, but eventually had to
resort to logging out how many boxes were in the map on the real input
to find the case where mine broke the map by changing the # of boxes.

I originally started recursing down the tree to find which boxes to push
This was originally flawed because I forgot to replace '.' where boxes
were moved out of. Once I realized that flaw, I tried a few things to
hack it in a recursive, DFS way, which worked on all but the real input.
But in that input, with a true triangular box array, my DFS hacks (e.g.
look at previous location to check if box) didn't work, so I had to go
back and make the search BFS so that at each depth, we can be sure that
checking depth-1 will determine whether the box's prior spot should be
replaced with a '.'.