Day 16 Part 2: Gahhhh that was frustrating
Having code that works by accident on the smaller example stinks.

My main issue was being too permissive with the candidates; just because
a neighbor has a lower score doesn't mean it's actually part of the path

To be part of the path, the distance has to be correct, which affected
both of the findPaths I tried in the last impl.