Day 18 Part 2: Easy as heck
Wow, this was the easiest Part 2 yet. I think this is the first day that
Part 2 has taken me less time than Part 1. For each obstacle, you just
check if there is a path to the end with all the previous obstacles. The
first one you find that doesn't have a path is the answer!

I even think this is close to optimal in terms of Big-N. Maybe there's a
slightly quicker way to tell by analyzing the grid for blocking paths
at every obstacle added instead of searching to the end for the entire
grid, or binary searching for the obstacle, but I don't care enough to
figure it out.

🍏 time python3 day18.py pt1 input.txt
['day18.py', 'pt1', 'input.txt']
answer: 296

________________________________________________________
Executed in   44.99 millis    fish           external
   usr time   25.14 millis   44.00 micros   25.10 millis
   sys time   12.08 millis  446.00 micros   11.63 millis

🍏 time python3 day18.py pt2 input.txt
['day18.py', 'pt2', 'input.txt']
Found the blocker: Coord(x=28, y=44)

________________________________________________________
Executed in   10.49 secs    fish           external
   usr time   10.46 secs   68.00 micros   10.46 secs
   sys time    0.03 secs  868.00 micros    0.03 secs