Day 19 Part 1: Recursion with a simple memoization required
This is clearly a recursive problem. Recursion: For <pattern>, if ∃towel
== prefix(pattern) AND recurse(rest of the pattern) = True, then pattern
can be created. Base case is the empty pattern is obviously true.

This recursion alone works for the test, but takes way too long for the
real input. But a simple memoization speeds this up enough to work:
🍏 time python3 day19.py pt1 input.txt
['day19.py', 'pt1', 'input.txt']
score: 365

________________________________________________________
Executed in  197.56 millis    fish           external
   usr time  174.62 millis   52.00 micros  174.56 millis
   sys time   16.97 millis  817.00 micros   16.15 millis