Day 19 Part 2: Same recursion, but different exit condition
This one didn't require many coding iterations, but did just a bit of
thinking to get both the iteration and memoization conditions. I had a
plan originally to store every possible way to get to a given pattern. I
realized that wasn't necessary; and that I could do a recursion where I
just add up all the possible options for the rest of the string based on
each of the towels that fit the current prefix.

Then the memoization is simply the value of how many ways there are to
make each of the suffix strings. It cleanly handles the strings that
have no way to create them because the total of the options starts at 0
and only adds more if one of the towels fits the prefix.

🍏 time python3 day19.py pt2 input.txt
['day19.py', 'pt2', 'input.txt']
score: 730121486795169

________________________________________________________
Executed in  405.70 millis    fish           external
   usr time  392.62 millis   59.00 micros  392.56 millis
   sys time   11.91 millis  837.00 micros   11.08 millis