Day 20 Part 1: Pathfinding + brute force
Uses the same BFS pathfinding search code from Day 18 to find the min
cost to the End. Then just bruteforces it by removing each obstacles and
finding the min cost to the end, and comparing that vs. the baseline.

There's a bunch of faster ways to do this:
1. Use the constraint that there's only a single path to the end to just
   iterate through the maze instead of BFS (I suspect this is only a
   small constant factor improvement)
2. Significantly faster: Once you know the distance for each item,
   iterate through again. At each location: Check the "other side" of
   the obstacles touching the square. If the distance on that "other
   side" is 100 less than the current distance, add 1 to the count.
   That's O(N) instead of my current approach, which I think is O(N^2).

But Imma spend my time on Part 2 instead of implementing that since this
brute-force was fast enough:
🍏 time python3 day20.py pt1 input.txt
['day20.py', 'pt1', 'input.txt']
total: 1296

________________________________________________________
Executed in   99.94 secs    fish           external
   usr time   99.68 secs   60.00 micros   99.68 secs
   sys time    0.26 secs  680.00 micros    0.25 secs