Day 20 Part 2: Finally got it working
Had a very similar approach that worked for finding the test example and
finding the
There are 4 cheats that save 74 picoseconds.
There are 3 cheats that save 76 picoseconds.
But had some subtly wrong logic in findCheats. The issue was that I was
"giving up" on cheats as soon as a cheat hit the path, but you have to
keep checking all the cheats through the entire maximum distance to find
every single cheat that can be reached from a given starting point on
the path.