Day 23 Part 1: Pretty basic graph manipulation
At first I thought mapping all the nodes to edges might be wasteful, but
then realized it was necessary and fairly straightforward to find all
the triples once you had the edges. The only mildly tricky bit was the
condition to add `other` to triples; at first I forgot the condition
that `other` needed to be in both `connection[0]` and `connection[1]`
edges, and after that forgot to check that `other` wasn't one of the
items already being considered, but after those brief challenges I got
it fairly quickly.

Could probably avoid the loop through `triples` by checking the `t`
condition before adding to triples, but it's already quick enough that
I don't care to clean it up

🍏 time python3 day23.py pt1 input.txt
['day23.py', 'pt1', 'input.txt']
count: 1419

________________________________________________________
Executed in   32.60 millis    fish           external
   usr time   25.11 millis   31.00 micros   25.08 millis
   sys time    6.78 millis  339.00 micros    6.44 millis