Day 23 Part 2: Works...but nondeterministically
This is some of the most confusing behavior I've seen. I wrote the
recursion correct fairly quickly (it got the answer right for the test
input). My intuition started from the approach for triples: go through
all the (pair0, pair1). For all edges[pair0]: e, we know pair0 is in
edges[e] since this is a bidirectional graph. So if pair1 is also in
edges[e], then we know there's an edge between all 3 of the nodes as
desired.

Then I thought "well, how can we find larger subsets on the same lines".
Similarly: assuming a given subset S of connected nodes, for arbitrary
other node n not in S, if S is a subset of edges[n], then n is also in
the same connected graph (analogous to how pair1 being in edges[e] means
there's a triple). With this inductive case, we just have to figure out
the base case, which is simply a single node which is always connected
to itself. And we can just keep track of the largest of the subsets.

Memoization was a bit tricky. Initially I tried to be conservative, with
a memoization key of (start, seen). But sets aren't hashable by default
and I got annoyed trying to make `seen` into a frozenset, so I figured
I'd give it a shot by just memoizing on `start`, assuming the largest
set including a given node would have to be the same regardless of the
status of the `seen` set.

Oddly enough, this gave me the right answer, but appears to not have
been a totally valid assumption, as the order of iteration through
`edges[start]` appears to affect the correctness, e.g.
🍏 python3 day23.py pt2 test.txt
['day23.py', 'pt2', 'test.txt']
largest: co,de,ka,ta
🍏 python3 day23.py pt2 test.txt
['day23.py', 'pt2', 'test.txt']
largest: ub,vc,wq
🍏 python3 day23.py pt2 test.txt
['day23.py', 'pt2', 'test.txt']
largest: co,de,ka,ta