WIP: Day 21 Part 2.
I'm not sure if the greedy algorithm I described in 5187edd won't work.
For example, ">>>A" is longer than "<>A", but the next iteration of
those would be "vAAA^A" vs. "v<<A>>A^A" respectively. So "only inspect
the shortest at each level" doesn't appear to be a valid greedy algo.