Day 21 Part 2: Holyyyy S, finally
This was the hardest problem yet for me. First I tried a brute-force
refactored implementation of 5187edd, but that clearly wasn't efficient
enough to work for the 25 iterations.

Then I had a decent idea for a greedy algorithm...but the greedy step
was incorrect.

Then I had the *correct* greedy algorithm correct (go all the way down
the tree of paths for a single component of a path, then cache the
results at each step to avoid recalculating the same thing a bunch of
times), but was missing the key fact that each path of a component needs
to start with an implicit "A" from the prior path at the next level of
depth, which made me think my greedy approach was wrong, and then spent
multiple nights debugging that.

But I got it, and that's nice.