Day 22 Part 2: A fun one with permutations
This was a fun one. I started off messing with the "mixing" & "pruning"
number evolution operations, since they all involve powers of 2, so
figured there might be some reverse-engineering-like problem as in D17.
But after messing around with that for some time I gave up and started
to test out how long brute-like approaches would take.

The first thing that came to my mind was the fact that since we are only
looking at the number % 10, all price differences will be in the range
[-9, 9], and the 19^4 permutations of these differences are actually not
THAT large of a number (130k). So my original approach was: for each of
these permutations, find the largest price in each of the seed # seqs.
But this seemed like it was going really slow (O(hours)) after trying it
so I tried to think of better ways to optimize. I realized I was
reprocessing each seed #s sequence of 2000 numbers 130k times, and that
there was probably a way to trade all of that reprocessing for space.

Finally, I realized I could just store the previous 4 numbers while
iterating through the sequence along with the price for the 4 numbers,
then add those to a centralized map for all of the seed-sequences, then
find the max of that centralized map. I had a minor issue of storing the
MAXIMUM price for a 4-tuple within sequence instead of just the first
occurrence, but after fixing that it ended up pretty quick:

🍏 time python3 day22.py pt2 input.txt
['day22.py', 'pt2', 'input.txt']
maxScore: 2018 - seq: (0, 4, -2, 2)

________________________________________________________
Executed in    1.91 secs    fish           external
   usr time    1.90 secs   46.00 micros    1.90 secs
   sys time    0.01 secs  455.00 micros    0.01 secs