Day 24 Part 1: A bit of pre-processing for a simple algo
Had to parse out the inputs and gates into a usable format, but from
there just did a BFS through the gates. At every step, we start with the
newly-found results, check if the gates associated with those results
can now be evaluated that weren't already evaluated, then add the new
results to a queue to check next time.

🍏 time python3 day24.py pt1 input.txt
['day24.py', 'pt1', 'input.txt']
51410244478064

________________________________________________________
Executed in   58.05 millis    fish           external
   usr time   25.16 millis   38.00 micros   25.12 millis
   sys time   15.59 millis  478.00 micros   15.11 millis