Day 3 Part 1
Pretty straightforward for me after seeing the example. It was clear
that you wanted to start with the largest number in the 10s place, then
find the largest number _after_ that to use as the 1s. I just fought
with Rust iterators to get a "pretty" (debatable...) implementation
instead of knocking out something quick and dirty with an inner loop.

🍏 time target/debug/cli 3 1
Day 3.1 output: XXXX

________________________________________________________
Executed in   11.15 millis    fish           external
   usr time    5.00 millis    0.07 millis    4.94 millis
   sys time    5.00 millis    1.13 millis    3.87 millis